run in highest-numbered display by default
SDL (and also Love, by inheritance) has a limitation in which you
can't determine the display where the mouse pointer is at startup.
So, instead of using the first display, we pick that the highest
numbered one, since it's usually the case that in a multimonitor
setup that is the external (and main) one.